+++ integration_tests/suite/test_phone_view.py
@@ -135,12 +135,6 @@ class TestPhoneView(BaseDirdIntegrationTest):
         )
         assert_that(result.status_code, equal_to((200)))
 
-    def test_gigaset_lookup_route(self):
-        result = self.get_lookup_gigaset_result(
-            self.profile, VALID_UUID, term='Alice', token=VALID_TOKEN_MAIN_TENANT
-        )
-        assert_that(result.status_code, equal_to((200)))
-
     def test_htek_lookup_route(self):
         result = self.get_lookup_htek_result(
             self.profile, VALID_UUID, term='Alice', token=VALID_TOKEN_MAIN_TENANT